geocoding-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai>=1.3.0
google-generativeai>=0.5.0
